libc: the bundled headers declare the C99 and POSIX types on every target
<stdio.h>, <stdlib.h>, <string.h>, <strings.h>, <time.h>, <unistd.h>,
<wchar.h> and <math.h> declared many library functions with the types of
an earlier dialect: `char *` for `void *` and for `const char *`, `int`
for `size_t`, `ssize_t` and `unsigned int`, `double` for `long double`.
A program that declares such a function with the standard's prototype
contradicts the header (C99 6.7p4), and a call converts its arguments to
the header's types: at edb82e4ff a 3 GiB `malloc` request reached the
library as 0xffffffffc0000000, and `fread`, `read` and `strspn` results
past 2^31 were truncated.

Each declaration now carries the C99 or POSIX type, spelled with the
target's `size_t`, `ssize_t`, `off_t` and `wchar_t`. On Windows `read`,
`write` and `getcwd` keep the CRT's `int` counts, which is what the
functions they bind take. `fabsl` and `ldexpl` are defined over `fabs`
and `ldexp`: c5 represents `long double` in binary64, so the `double`
functions compute them exactly, and no platform entry point receives a
long double in a format its ABI does not use. <time.h> includes
<stddef.h> for `size_t` (C99 7.23.1p2). The definitions libc/lib supplies
for Windows (`snprintf`, `vsnprintf`, `strndup`, `memrchr`) follow their
declarations.

bound_import_arg_narrowing.c relied on `memcmp` and `memset` narrowing a
64-bit count to `int`; it now checks that `strncmp`'s `size_t` count keeps
its high bits while a user callee's `int` parameter still narrows.
extern_in_function.c declared `strchr` with a `char *` parameter, and a
string-extension test subtracted a `char *` from the `void *` `memrchr`
returns; both now use the standard types. The `fabsl` linker test checks
the definitions in SSA instead of an import of `fabs`, and a codegen test
checks that `malloc`'s argument reaches the call unextended on every
target.

Snapshots: the fixtures that pass or receive these values drop the 32-bit
sign extension (`sxtw` and `movslq`, 23 each; SSA `Extend` and its `shl`),
a constant `size_t` argument no longer leaves the `n << 32` intermediate
of that extension behind (80 `Imm`), c4 passes its sizes as 64-bit
`movq`, and every unit including <math.h> numbers its functions two
entries later for the `fabsl` / `ldexpl` definitions.

+++ libc/include/wchar.h
@@ -195,11 +195,11 @@ double wcstod(const wchar_t *nptr, wchar_t **endptr);
 wchar_t *wcstok(wchar_t *s, const wchar_t *delim, wchar_t **ptr);
 
 // Restartable multibyte / wide conversion (C99 7.24.6.3).
-unsigned long mbrtowc(wchar_t *pwc, const char *s, unsigned long n, mbstate_t *ps);
-unsigned long wcrtomb(char *s, wchar_t wc, mbstate_t *ps);
-unsigned long wcsrtombs(char *dst, const wchar_t **src, unsigned long len, mbstate_t *ps);
+size_t mbrtowc(wchar_t *pwc, const char *s, size_t n, mbstate_t *ps);
+size_t wcrtomb(char *s, wchar_t wc, mbstate_t *ps);
+size_t wcsrtombs(char *dst, const wchar_t **src, size_t len, mbstate_t *ps);
 // C95 7.24.6.4.1: the multibyte->wide reverse of wcsrtombs.
-unsigned long mbsrtowcs(wchar_t *dst, const char **src, unsigned long len, mbstate_t *ps);
+size_t mbsrtowcs(wchar_t *dst, const char **src, size_t len, mbstate_t *ps);
 
 // Single-byte / wide-character conversion (C99 7.24.6.1).
 wint_t btowc(int c);
